refactor(backend): split commands into files
