+++ src-tauri/src/blackhole/mod.rs
@@ -1,6 +1,7 @@
-mod logic;
+mod scan_blackhole_stream;
+mod import_blackhole;
 
-pub use logic::{BlackholeFolder};
+pub use scan_blackhole_stream::BlackholeFolder;
 
 #[tauri::command]
 pub async fn scan_blackhole_stream(
@@ -8,7 +9,7 @@ pub async fn scan_blackhole_stream(
     root_path: String,
     dest_path: String,
 ) -> Result<Vec<BlackholeFolder>, String> {
-    logic::scan_blackhole_stream(window, root_path, dest_path).await
+    scan_blackhole_stream::scan_blackhole_stream(window, root_path, dest_path).await
 }
 
 #[tauri::command]
@@ -17,5 +18,5 @@ pub async fn import_blackhole(
     dest_path: String,
     cut: bool,
 ) -> Result<(), String> {
-    logic::import_blackhole(files, dest_path, cut).await
+    import_blackhole::import_blackhole(files, dest_path, cut).await
 }
