+++ src-tauri/src/duplicate/delete_files.rs
@@ -1,0 +1,6 @@
+pub fn delete_files(paths: Vec<String>) -> Result<(), String> {
+    for p in paths {
+        std::fs::remove_file(&p).map_err(|e| e.to_string())?;
+    }
+    Ok(())
+}
