+++ src-tauri/src/importer/external_device.rs
@@ -1,0 +1,8 @@
+use serde::Serialize;
+
+#[derive(Serialize)]
+pub struct ExternalDevice {
+    pub name: String,
+    pub path: String,
+    pub total: u64,
+}
