+++ src-tauri/src/importer/mod.rs
@@ -1,18 +1,21 @@
-mod logic;
+mod external_device;
+mod list_external_devices;
+mod list_all_disks;
+mod import_device;
 
-pub use logic::ExternalDevice;
+pub use external_device::ExternalDevice;
 
 #[tauri::command]
 pub fn list_external_devices() -> Result<Vec<ExternalDevice>, String> {
-    logic::list_external_devices()
+    list_external_devices::list_external_devices()
 }
 
 #[tauri::command]
 pub fn list_all_disks() -> Result<Vec<ExternalDevice>, String> {
-    logic::list_all_disks()
+    list_all_disks::list_all_disks()
 }
 
 #[tauri::command]
 pub async fn import_device(device_path: String, dest_path: String) -> Result<(), String> {
-    logic::import_device(device_path, dest_path).await
+    import_device::import_device(device_path, dest_path).await
 }
