+++ src-tauri/src/sort/mod.rs
@@ -1,11 +1,13 @@
-mod logic;
+mod scan_images;
+mod find_images;
+mod sort_images;
 
 #[tauri::command]
 pub fn find_images(path: String) -> Result<Vec<String>, String> {
-    logic::find_images(path)
+    find_images::find_images(path)
 }
 
 #[tauri::command]
 pub fn sort_images(path: String) -> Result<(), String> {
-    logic::sort_images(path)
+    sort_images::sort_images(path)
 }
