+++ src-tauri/src/training/training_entry.rs
@@ -1,0 +1,8 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize, Clone)]
+pub struct TrainingEntry {
+    pub tag: String,
+    pub path: String,
+    pub delete: Option<bool>,
+}
